Implements timer and external tick
